// frontend/src/app/models/tutorial.model.ts
export class Tutorial {
  id?: any;
  name?: string;
  provider?: string;
  photo?: string;
  idIndex?: string;
  sex?: string;
  age?: number;
  phone?: string;
  order?: number;
  processByName?: string;
  processByEmail?: string;
  date?: string;
  subtotal?: number;
  covered?: number;
  published?: boolean;
}